Solution for C: Task 5

+++ script.sh
@@ -1,38 +1,49 @@
-#!/usr/bin/env bash
+### Done Task 5
 
-### Done Task 4
-for file in $*
-do
-	gcc $file -o file.out
-	echo "Running $file on test cases"
-	for i in 1 2 3
+language=$1
+files=${*:2}
+echo "The language passed as argument is $language"
+
+if [[ "$language" != "C" ]]
+then
+	echo "Invalid language"
+
+else
+
+	for file in $files
 	do
-		echo "Running test case $i:"
-		./file.out < input$i.txt > user_output$i.txt 2> $file_error$i.txt
-		exit_code=$?
+		gcc $file -o file.out
+		echo "Running $file on test cases"
+		for i in 1 2 3
+		do
+			echo "Running test case $i:"
+			./file.out < input$i.txt > user_output$i.txt 2> $file_error$i.txt
+			exit_code=$?
 
-		if [[ $exit_code -eq 0 ]]
-		then
-			echo "Code ran without run-time errors"
-
-			passed=$(diff output$i.txt user_output$i.txt)
-			if [[ "$passed" == "" ]]
+			if [[ $exit_code -eq 0 ]]
 			then
-				echo "Test case $i passed"
+				echo "Code ran without run-time errors"
+
+				passed=$(diff output$i.txt user_output$i.txt)
+				if [[ "$passed" == "" ]]
+				then
+					echo "Test case $i passed"
+				else
+					echo "Test case $i failed"
+					echo "User output:"
+					cat user_output$i.txt
+					echo "Expected output:"
+					cat output$i.txt
+				fi
+
 			else
-				echo "Test case $i failed"
-				echo "User output:"
-				cat user_output$i.txt
-				echo "Expected output:"
-				cat output$i.txt
+				echo "Code has run-time errors"
+				cat $file_error$i.txt
 			fi
 
-		else
-			echo "Code has run-time errors"
-			cat $file_error$i.txt
-		fi
-
-		echo "------------------------------------------"
+			echo "------------------------------------------"
+		done
+		rm ./file.out
 	done
-	rm ./file.out
-done
+
+fi
